Force float32 casting for mixed precision stability
* Appended `dtype="float32"` to the final output Dense layer (`Normalized_Feedrate_Output`) in `train_bi_lstm.py` and `plan_B/train_bi_lstm.py`.
* This explicit casting prevents catastrophic numerical underflow (`NaN` gradients) when calculating the MAPE loss function during hardware-accelerated `mixed_float16` Tensor Core executions.

+++ train_bi_lstm.py
@@ -42,7 +42,7 @@ def build_bilstm_model(input_shape: Tuple[int, int], learning_rate: float = 1e-3
     d = layers.Dense(128, activation="relu")(merged)
     d = layers.Dropout(0.2)(d)
     d = layers.Dense(64, activation="relu")(d)
-    outputs = layers.Dense(1, activation="linear", name="Normalized_Feedrate_Output")(d)
+    outputs = layers.Dense(1, activation="linear", name="Normalized_Feedrate_Output", dtype="float32")(d)
 
     model = models.Model(inputs=inputs, outputs=outputs, name="CNC_Kinematics_BiLSTM_V2")
